Moving query to pcie queries
