+++ src/runtime_src/core/edge/user/device_linux.cpp
@@ -5,7 +5,6 @@
 #include "xrt.h"
 #include "zynq_dev.h"
 #include "aie_sys_parser.h"
-#include "smi.h"
 
 #include "core/common/debug_ip.h"
 #include "core/common/query_requests.h"
@@ -666,29 +665,6 @@ struct aim_counter
   }
 };
 
-struct xrt_smi_config
-{
-  using result_type = std::any; 
-
-  static result_type
-  get(const xrt_core::device* device, key_type key, const std::any& reqType)
-  {
-    if (key != xrt_key_type::xrt_smi_config)
-      throw xrt_core::query::no_such_key(key, "Not implemented");
-    std::string xrt_smi_config;
-    const auto xrt_smi_config_type = std::any_cast<xrt_core::query::xrt_smi_config::type>(reqType);
-    switch (xrt_smi_config_type) {
-    case xrt_core::query::xrt_smi_config::type::options_config:
-      xrt_smi_config = xrt_core::shim::smi::get_smi_config();
-      break;
-    default:
-      throw xrt_core::query::no_such_key(key, "Not implemented");
-    }
-
-    return xrt_smi_config;
-  }
-};
-
 struct am_counter
 {
   using result_type = query::am_counter::result_type;
@@ -1079,7 +1055,6 @@ initialize_query_table()
   emplace_func0_request<query::instance,                instance>();
   emplace_func0_request<query::xclbin_slots,            xclbin_slots>();
 
-  emplace_func4_request<query::xrt_smi_config,          xrt_smi_config>();
   emplace_func4_request<query::aim_counter,             aim_counter>();
   emplace_func4_request<query::am_counter,              am_counter>();
   emplace_func4_request<query::asm_counter,             asm_counter>();
