+++ src/runtime_src/core/pcie/linux/device_linux.cpp
@@ -17,6 +17,7 @@
 #include "core/include/xdp/spc.h"
 #include "core/pcie/driver/linux/include/mgmt-ioctl.h"
 
+#include "smi.h"
 #include "pcidev.h"
 #include "xrt.h"
 
@@ -906,6 +907,30 @@ struct am_counter
 
 };
 
+struct xrt_smi_config
+{
+  using result_type = std::any; 
+
+  static result_type
+  get(const xrt_core::device* device, key_type key, const std::any& reqType)
+  {
+    if (key != xrt_key_type::xrt_smi_config)
+      throw xrt_core::query::no_such_key(key, "Not implemented");
+    std::string xrt_smi_config;
+    const auto xrt_smi_config_type = std::any_cast<xrt_core::query::xrt_smi_config::type>(reqType);
+    switch (xrt_smi_config_type) {
+    case xrt_core::query::xrt_smi_config::type::options_config:
+      xrt_smi_config = xrt_core::shim::smi::get_smi_config();
+      break;
+    default:
+      throw xrt_core::query::no_such_key(key, "Not implemented");
+    }
+
+    return xrt_smi_config;
+  }
+};
+
+
 
 /* ASM counter values
  * In PCIe Linux, access the sysfs file for ASM to retrieve the ASM counter values
@@ -1457,6 +1482,7 @@ initialize_query_table()
   emplace_func4_request<query::aim_counter,                    aim_counter>();
   emplace_func4_request<query::am_counter,                     am_counter>();
   emplace_func4_request<query::asm_counter,                    asm_counter>();
+  emplace_func4_request<query::xrt_smi_config,                 xrt_smi_config>();
   emplace_func4_request<query::lapc_status,                    lapc_status>();
   emplace_func4_request<query::spc_status,                     spc_status>();
   emplace_func4_request<query::accel_deadlock_status,          accel_deadlock_status>();
